requirements.txt
dnspython==2.7.0
paho-mqtt==2.1.0
pymongo==4.11.1
